Externalize the function deserializeCodeLines
